backend/telemetry_service/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
pydantic
PyJWT